Refactoring into core folder
- Create folder structure and code to appropriate named files
- Remove unnecessary/old comments
- Remove __precompile__()
- Instead of importing push! and reset!, prepend function definition with module
- export @rangecheck
- moved FeasibilityFormNLS of a FeasibilityResidual to model-interaction.jl
- moved hess functions of a FeasibilityResidual{LLSModel} to model-interaction.jl
