+++ src/core/nlp/utils.jl
@@ -1,6 +1,6 @@
-export coo_prod!, coo_sym_prod!,
-       @default_counters, @default_nlscounters,
-       DimensionError, @lencheck
+export coo_prod!, coo_sym_prod!
+export @default_counters
+export DimensionError, @lencheck, @rangecheck
 
 struct DimensionError <: Exception
   name :: Union{Symbol,String}
@@ -89,22 +89,6 @@ macro default_counters(Model, inner)
                      reset_data!(nlp)
                    end
                   ))
-  push!(ex.args, :(NLPModels.increment!(nlp :: $(esc(Model)), s :: Symbol) = increment!(nlp.$inner, s)))
-  push!(ex.args, :(NLPModels.decrement!(nlp :: $(esc(Model)), s :: Symbol) = decrement!(nlp.$inner, s)))
-  ex
-end
-
-"""
-    @default_nlscounters Model inner
-
-Define functions relating NLS counters of `Model` to NLS counters of `Model.inner`.
-"""
-macro default_nlscounters(Model, inner)
-  ex = Expr(:block)
-  for foo in fieldnames(NLSCounters) ∪ [:sum_counters, :reset!]
-    foo == :counters && continue
-    push!(ex.args, :(NLPModels.$foo(nlp :: $(esc(Model))) = $foo(nlp.$inner)))
-  end
   push!(ex.args, :(NLPModels.increment!(nlp :: $(esc(Model)), s :: Symbol) = increment!(nlp.$inner, s)))
   push!(ex.args, :(NLPModels.decrement!(nlp :: $(esc(Model)), s :: Symbol) = decrement!(nlp.$inner, s)))
   ex
