+++ src/core/nls/api.jl
@@ -1,102 +1,9 @@
-export AbstractNLSModel, nls_meta, NLSCounters, reset!,
-       residual, residual!, jac_residual, jac_structure_residual, jac_structure_residual!,
-       jac_coord_residual!, jac_coord_residual, jprod_residual, jprod_residual!,
-       jtprod_residual, jtprod_residual!, jac_op_residual, jac_op_residual!,
-       hess_residual, hess_structure_residual, hess_structure_residual!,
-       hess_coord_residual!, hess_coord_residual, jth_hess_residual,
-       hprod_residual, hprod_residual!, hess_op_residual, hess_op_residual!
-
-abstract type AbstractNLSModel <: AbstractNLPModel end
-
-mutable struct NLSCounters
-  counters :: Counters
-  neval_residual        :: Int
-  neval_jac_residual    :: Int
-  neval_jprod_residual  :: Int
-  neval_jtprod_residual :: Int
-  neval_hess_residual   :: Int
-  neval_jhess_residual  :: Int
-  neval_hprod_residual  :: Int
-
-  function NLSCounters()
-    return new(Counters(), 0, 0, 0, 0, 0, 0, 0)
-  end
-end
-
-import Base.getproperty, Base.setproperty!
-function getproperty(c :: NLSCounters, f :: Symbol)
-  if f in fieldnames(Counters)
-    getfield(c.counters, f)
-  else
-    getfield(c, f)
-  end
-end
-
-function setproperty!(c :: NLSCounters, f :: Symbol, x)
-  if f in fieldnames(Counters)
-    setfield!(c.counters, f, x)
-  else
-    setfield!(c, f, x)
-  end
-end
-
-function sum_counters(c :: NLSCounters)
-  s = sum_counters(c.counters)
-  for field in fieldnames(NLSCounters)
-    field == :counters && continue
-    s += getfield(c, field)
-  end
-  return s
-end
-sum_counters(nls :: AbstractNLSModel) = sum_counters(nls.counters)
-
-for counter in fieldnames(NLSCounters)
-  counter == :counters && continue
-  @eval begin
-    """
-    $($counter)(nlp)
-
-    Get the number of `$(split("$($counter)", "_")[2])` evaluations.
-    """
-    $counter(nls :: AbstractNLSModel) = nls.counters.$counter
-    export $counter
-  end
-end
-
-for counter in fieldnames(Counters)
-  @eval begin
-    $counter(nls :: AbstractNLSModel) = nls.counters.counters.$counter
-    export $counter
-  end
-end
-
-import NLPModels: reset!
-
-function reset!(nls :: AbstractNLSModel)
-  reset!(nls.counters)
-  return nls
-end
-
-function reset!(nls_counters :: NLSCounters)
-  for f in fieldnames(NLSCounters)
-    f == :counters && continue
-    setfield!(nls_counters, f, 0)
-  end
-  reset!(nls_counters.counters)
-  return nls_counters
-end
-
-include("nls_meta.jl")
-
-nls_meta(nls :: AbstractNLSModel) = nls.nls_meta
-
-include("autodiff_nlsmodel.jl")
-include("lls_model.jl")
-include("feasibility_residual.jl")
-
-# min ½‖F(x)‖²
-
-# TODO: Add a regularization ½λ²‖V(x)‖²
+export residual, residual!, jac_residual, jac_structure_residual, jac_structure_residual!
+export jac_coord_residual!, jac_coord_residual, jprod_residual, jprod_residual!
+export jtprod_residual, jtprod_residual!, jac_op_residual, jac_op_residual!
+export hess_residual, hess_structure_residual, hess_structure_residual!
+export hess_coord_residual!, hess_coord_residual, jth_hess_residual
+export hprod_residual, hprod_residual!, hess_op_residual, hess_op_residual!
 
 """
     Fx = residual(nls, x)
