+++ src/feasibility_form_nls.jl
@@ -65,33 +65,6 @@ function FeasibilityFormNLS(nls :: AbstractNLSModel; name="$(nls.meta.name)-ffnl
   return nlp
 end
 
-function FeasibilityFormNLS(nls :: FeasibilityResidual; name="$(nls.meta.name)-ffnls")
-  meta = nls.nlp.meta
-  nequ = meta.ncon
-  nvar = meta.nvar + nequ
-  ncon = meta.ncon
-  nnzj = meta.nnzj + nequ
-  nnzh = meta.nnzh + nequ
-  meta = NLPModelMeta(nvar, x0=[meta.x0; zeros(nequ)],
-                      lvar=[meta.lvar; fill(-Inf, nequ)],
-                      uvar=[meta.uvar; fill( Inf, nequ)],
-                      ncon=ncon,
-                      lcon=meta.lcon,
-                      ucon=meta.ucon,
-                      y0=meta.y0,
-                      lin=meta.lin,
-                      nln=meta.nln,
-                      nnzj=nnzj, nnzh=nnzh,
-                      name=name
-                     )
-  nls_meta = NLSMeta(nequ, nvar, x0=[meta.x0; zeros(nequ)], nnzj=nequ, nnzh=0)
-
-  nlp = FeasibilityFormNLS{FeasibilityResidual}(meta, nls_meta, nls, NLSCounters())
-  finalizer(nlp -> finalize(nlp.internal), nlp)
-
-  return nlp
-end
-
 function obj(nlp :: FeasibilityFormNLS, x :: AbstractVector)
   @lencheck nlp.meta.nvar x
   increment!(nlp, :neval_obj)
@@ -213,14 +186,6 @@ function hess_structure!(nlp :: FeasibilityFormNLS, rows :: AbstractVector{<: In
   return rows, cols
 end
 
-function hess_structure!(nlp :: FeasibilityFormNLS{LLSModel}, rows :: AbstractVector{<: Integer}, cols :: AbstractVector{<: Integer})
-  @lencheck nlp.meta.nnzh rows cols
-  n, ne = nlp.internal.meta.nvar, nlp.internal.nls_meta.nequ
-  rows .= n+1:n+ne
-  cols .= n+1:n+ne
-  return rows, cols
-end
-
 function hess_coord!(nlp :: FeasibilityFormNLS, xr :: AbstractVector, y :: AbstractVector, vals :: AbstractVector;
                      obj_weight :: Float64=1.0)
   @lencheck nlp.meta.nvar xr
@@ -240,16 +205,6 @@ function hess_coord!(nlp :: FeasibilityFormNLS, xr :: AbstractVector, y :: Abstr
     @views hess_coord!(nlp.internal, x, y2, vals[I], obj_weight=0.0)
   end
   vals[nnzhF+nnzhc+1:nnzhF+nnzhc+ne] .= obj_weight
-  return vals
-end
-
-function hess_coord!(nlp :: FeasibilityFormNLS{LLSModel}, xr :: AbstractVector, y :: AbstractVector, vals :: AbstractVector;
-                     obj_weight :: Float64=1.0)
-  @lencheck nlp.meta.nvar xr
-  @lencheck nlp.meta.ncon y
-  @lencheck nlp.meta.nnzh vals
-  increment!(nlp, :neval_hess)
-  vals .= obj_weight
   return vals
 end
 
@@ -290,14 +245,14 @@ function hprod!(nlp :: FeasibilityFormNLS, xr :: AbstractVector, y :: AbstractVe
   return hv
 end
 
-function ghjvprod!(nlp :: FeasibilityFormNLS, x :: AbstractVector, g :: AbstractVector, v :: AbstractVector, gHv :: AbstractVector) 
+function ghjvprod!(nlp :: FeasibilityFormNLS, x :: AbstractVector, g :: AbstractVector, v :: AbstractVector, gHv :: AbstractVector)
   @lencheck nlp.meta.nvar x g v
   @lencheck nlp.meta.ncon gHv
   increment!(nlp, :neval_hprod)
   n, m, ne = nlp.internal.meta.nvar, nlp.internal.meta.ncon, nlp.internal.nls_meta.nequ
   IF = 1:ne
   Ic = ne+1:ne+m
-  gHv[IF] .= [dot(g[1:n], hprod_residual(nlp.internal, x[1:n], j, v[1:n])) for j in IF]  
+  gHv[IF] .= [dot(g[1:n], hprod_residual(nlp.internal, x[1:n], j, v[1:n])) for j in IF]
   if m > 0
     @views ghjvprod!(nlp.internal, x[1:n], g[1:n], v[1:n], gHv[Ic])
   end
